Create terraform files to provision a GCS bucket
We added the terraform files to specify Google as the provider
and provision a GCS bucket resource. The variables are stored in a
`terraform.tfvars` that is ignored when committed. Since the key
is in json, we have also ignored that in the commit. Ideally, the
key is to be stored in a secrets manager or using Google's
Workload Identity Federation.